Rename CLI files
