+++ .vscode/settings.json
@@ -4,6 +4,8 @@
     "jenkins.pipeline.linter.connector.url": "https://jenkins.symboxtra.dynu.net/pipeline-model-converter/validate",
     "jenkins.pipeline.linter.connector.user": "symboxtra",
     "files.associations": {
-        "xstring": "cpp"
+        "xstring": "cpp",
+        "deque": "cpp",
+        "vector": "cpp"
     }
 }
